+++ template-parts/footer/site-info.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.1.0
  */
 
 
@@ -15,7 +15,7 @@
 
 // Helper variables
 
-	$site_info_text = trim( get_theme_mod( 'texts_site_info' ) );
+	$site_info_text = trim( (string) get_theme_mod( 'texts_site_info' ) );
 
 
 // Requirements check
@@ -50,6 +50,10 @@
 					'<a href="' . esc_url( wp_get_theme( 'modern' )->get( 'ThemeURI' ) ) . '"><strong>' . wp_get_theme( 'modern' )->get( 'Name' ) . '</strong></a>',
 					'<a href="' . esc_url( __( 'https://wordpress.org/', 'modern' ) ) . '">WordPress</a>'
 				);
+
+				if ( function_exists( 'the_privacy_policy_link' ) ) {
+					the_privacy_policy_link( '<span class="sep"> | </span>' );
+				}
 
 				?>
 				<span class="sep"> | </span>